+++ tests/test_db.py
@@ -4,7 +4,6 @@ import unittest
 import doctest
 
 import tsprocess.database as database
-from tsprocess.database import DataBase
 
 class TestDataBase(unittest.TestCase):
     def setUp(self):
@@ -12,13 +11,13 @@ class TestDataBase(unittest.TestCase):
         self.csize = 10
 
     def test_create_database_singleton(self):
-        db_1 = DataBase(self.dbname, self.csize)
-        db_2 = DataBase(self.dbname, self.csize)
+        db_1 = database.DataBase(self.dbname, self.csize)
+        db_2 = database.DataBase(self.dbname, self.csize)
 
         self.assertEqual(db_1,db_2)
     
     def test_set_and_get_value(self):
-        db_3 = DataBase('mytest',10)
+        db_3 = database.DataBase('mytest',10)
         db_3.set_value('x1',100)
         self.assertEqual(db_3.get_value('x1'),100)
 
@@ -38,9 +37,6 @@ class TestDataBase(unittest.TestCase):
         tests.addTests(doctest.DocTestSuite(database))
         return tests
         
-
-
-
 
 suite = unittest.TestSuite()
 suite.addTest(doctest.DocTestSuite(database))