+++ tests/test_doctest.py
@@ -8,6 +8,7 @@ import tsprocess.ts_utils as ts_utils
 import tsprocess.database as database
 import tsprocess.incident as incident
 import tsprocess.timeseries as timeseries
+import tsprocess.ts_utils as ts_plot_utils
 import tsprocess.seismicsource as seismicsource
 
 
@@ -22,6 +23,7 @@ def test_doctest_suit():
     test_suit.addTest(doctest.DocTestSuite(database))
     test_suit.addTest(doctest.DocTestSuite(incident))
     test_suit.addTest(doctest.DocTestSuite(timeseries))
+    test_suit.addTest(doctest.DocTestSuite(ts_plot_utils))
     test_suit.addTest(doctest.DocTestSuite(seismicsource))
     
     # set runner